Implemented a renewed Layout system.
The layout system now uses so called TemplateEngines. A Template Engine is an engine that implements \FuzeWorks\TemplateEngine, and is a standardized way of parsing template files.
3 template engines are present by default: PHP Parser, JSON parser and the Smarty Template engine.
Based on the file extension the manager should be able to decide which engine to use. New engines can be added by using $this->layout->registerEngine();. This function should be called on the layoutLoadEngineEvent.
Also Smarty has been removed from the system and is now a composer dependency.